Cargill Prototype With Right Arrow for Analytics
This is one of the prototypes from me and my team, Prime Collaborators, for Accenture's L2L program. This version of the K.E.Y.S. HTML code has a right arrow which brings you to a page with a timeline for how to improve animal feed, which was our example product. The code was generated using Gemini, ChatGPT, and Copilot.