Cargill Prototype - First Edition
This is one of the first prototypes from me and my team, Prime Collaborators, for Accenture's L2L program. This version of the HTML code started off with the title "Automation System For Keywords Activated" and only had the revolving lines of code which was an example of the code connecting to APIs and public information systems for more discussion based social medias like Twitter (X), Threads, etc. The code was generated using Gemini, ChatGPT, and Copilot.